logEvent gives non-compacted events to DelightStreamingConnector
The objective is to speed up logEvent so that events are ingested fast
enough and the queue capacity
`spark.scheduler.listenerbus.eventqueue.capacity` is not reached.

The events are now compacted at the very end, just before being sent
over the network.

This change yields a 5x to 20x speedup.