Add a Delight version in the Spark config
Since the version of the JAR is fixed to `latest-SNAPSHOT` in order to
provide continuous upgrades to Delight users, we set a version in Spark
config `spark.delight.version`. The Spark config is accessible to
Delight backend, so we'll be able to leverage the agent version in the
dashboard, or at least inspect it when debugging.